Session 1/2. Exposing API/requirements.txt
Flask==2.0.1
flask_cors==3.0.10
Flask_RESTful==0.3.9
flask_restx==0.5.1
h5py==3.1.0
matplotlib==3.5.1
numpy==1.21.0
opencv_python_headless
Pillow==9.0.1
requests==2.27.1
scikit_image==0.18.1
scipy==1.7.3
tensorflow==2.7.2
Werkzeug==2.0.3
jupyter
gunicorn